requirements.txt
pandas>=2.0
requests>=2.28
python-dotenv>=1.0
jinja2>=3.1
tqdm>=4.65
finance-datareader>=0.9.50
pykrx>=1.0.40
exchange-calendars>=4.5
pyarrow>=14.0
scikit-learn>=1.3.0
numpy>=1.24
